Add files via upload
The count_char function counts the number of times a character occurs in a string. My other function has the same properties. I created it for user input. With Python's open command, I opened my text file containing the text and found the characters in it with the for loop with the help of my function. I used the upper command to avoid mixing upper- and lower-case letters.